chore: Move article model to /models
